+++ clase6/recu.py
@@ -1,0 +1,11 @@
+#Recursividad
+
+def fibo(n):
+    if n > 1:
+        return fibo(n-1) + fibo(n-2)
+    elif n==1 or n == 0:
+        return 1
+    elif n < 0:
+        print('Valor inválido')
+
+print(fibo(6))